Add ability to jump to and to delete buffers
Using ctrl-o you can jump (:drop) to the buffer if it's already open.
Using ctrl-d you can delete the selected buffer(s)

Add the option g:fzf_preview_buffers_jump (default 0) to drop to the
buffer by default. If set to 1 the behavior of ctrl-o will be to open
the buffer in the current window instead.